Enforcing unique sites. Modified file layout.

+++ src/components/StoredPasswordsTable/utils.js
@@ -1,57 +1,34 @@
-import { 
+import {
+  addDoc,
   collection, 
   deleteDoc,
   doc,
-  getDocs
+  getDoc,
+  serverTimestamp
 } from 'firebase/firestore';
-import * as CryptoJS from 'crypto-js';
 
 import { store } from '../../utils';
-import { 
-  getAESsecret,
-  UserSessionException, 
-  UserSecretNotFoundException 
+import {
+  DocumentNotFoundException
 } from '../common/';
-
-
-export const getStoredRecords = async () => {
-  const db = store;
-  const records = [];
-
-  const uid = sessionStorage.getItem('User ID');
-    
-  if (!uid) {
-    const errMsg = 'User ID not found in session storage';
-    throw new UserSessionException(errMsg);
-  }
-
-  const userSecret = getAESsecret();
-
-  if (!userSecret) {
-    const errMsg = 'User Client Side Secret not found in local storage';
-    throw new UserSecretNotFoundException(errMsg);
-  }
-
-  const userCollection = collection(db, 'records', `${uid}`, 'current');
-  const querySnapshot = await getDocs(userCollection);
-  
-  querySnapshot.forEach((doc) => {
-    let data = doc.data();
-    data['id'] = doc.id;
-
-    // Decrypt the passwords
-    let decryptedPwd = CryptoJS.AES.decrypt(data['password'], userSecret).toString(CryptoJS.enc.Utf8);
-    data['password'] = String(decryptedPwd);
-
-    records.push(data); 
-  });
-
-  return records;
-}
 
 export const deleteRecordEntry = async (id) => {
   const db = store;
-  const taskDocRef = doc(db, 'records', id);
-  
+  const uid =  sessionStorage.getItem('User ID');
+  const taskDocRef = doc(db, 'records', `${uid}`, 'current', id);
+
+  // Copy the document and move to deleted entry before deleting from current passwords
+  const docSnap = await getDoc(taskDocRef);
+
+  if (!docSnap.exists()) {
+    const errMsg = `Document [${id}] not found in store`;
+    throw new DocumentNotFoundException(errMsg);
+  }
+
+  let deletedData = docSnap.data();
+  deletedData['deleteTime'] = serverTimestamp();
+
+  await addDoc(collection(db, 'records', `${uid}`, 'deleted'), deletedData);
   await deleteDoc(taskDocRef);
 }
+
